pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Second project on Hexlet python course"
authors = ["Viktor Shcherbakov <evil.logitech@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
    "License :: Freeware",
    "Natural Language :: English"
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipykernel = "^6.21.3"
flake8-cognitive-complexity = "^0.1.0"
pytest-cov = "^4.0.0"
pytest = "7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
